requirements.txt
Django>=5.2.4
whitenoise
gunicorn
djangorestframework
pandas
numpy 